+++ bee-runtime/src/lib.rs
@@ -1,0 +1,12 @@
+// Copyright 2020 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! Runtime components and utilities for the bee node framework
+
+#![deny(missing_docs, warnings)]
+
+pub mod event;
+pub mod node;
+pub mod resource;
+pub mod shutdown_stream;
+pub mod worker;