+++ bee-runtime/tests/event.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_common::event::Bus;
+use bee_runtime::event::Bus;
 
 struct Foo;