+++ bee-common/bee-common/CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Removed `event::Bus` in favour of its inclusion in `bee-runtime`
+- Removed `shutdown_stream::ShutdownStream` in favour of its inclusion in `bee-runtime`
+- Removed `shutdown::Shutdown`
+- Removed `worker::Worker`
+
 ### Fixed
 
 ### Security -->